# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentStatus = typeof StudentStatus[keyof typeof StudentStatus];


export const StudentStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface Student {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  level: string;
  progress: number;
  attendanceRate: number;
  status: StudentStatus;
  joinedAt: string;
}

export type StudentInputStatus = typeof StudentInputStatus[keyof typeof StudentInputStatus];


export const StudentInputStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface StudentInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  level: string;
  status?: StudentInputStatus;
}

export type StudentUpdateStatus = typeof StudentUpdateStatus[keyof typeof StudentUpdateStatus];


export const StudentUpdateStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface StudentUpdate {
  /** @minLength 2 */
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  level?: string;
  status?: StudentUpdateStatus;
}

export type TeacherStatus = typeof TeacherStatus[keyof typeof TeacherStatus];


export const TeacherStatus = {
  active: 'active',
  on_leave: 'on_leave',
} as const;

export interface Teacher {
  id: number;
  name: string;
  email: string;
  specialty: string;
  studentsCount: number;
  status: TeacherStatus;
}

export type TeacherInputStatus = typeof TeacherInputStatus[keyof typeof TeacherInputStatus];


export const TeacherInputStatus = {
  active: 'active',
  on_leave: 'on_leave',
} as const;

export interface TeacherInput {
  /** @minLength 2 */
  name: string;
  email: string;
  specialty: string;
  status?: TeacherInputStatus;
}

export type TeacherUpdateStatus = typeof TeacherUpdateStatus[keyof typeof TeacherUpdateStatus];


export const TeacherUpdateStatus = {
  active: 'active',
  on_leave: 'on_leave',
} as const;

export interface TeacherUpdate {
  /** @minLength 2 */
  name?: string;
  email?: string;
  specialty?: string;
  status?: TeacherUpdateStatus;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  present: 'present',
  absent: 'absent',
  late: 'late',
  excused: 'excused',
} as const;

export interface AttendanceRecord {
  id: number;
  studentId: number;
  studentName: string;
  date: string;
  status: AttendanceRecordStatus;
  /** @nullable */
  note?: string | null;
}

export type AttendanceInputStatus = typeof AttendanceInputStatus[keyof typeof AttendanceInputStatus];


export const AttendanceInputStatus = {
  present: 'present',
  absent: 'absent',
  late: 'late',
  excused: 'excused',
} as const;

export interface AttendanceInput {
  studentId: number;
  date: string;
  status: AttendanceInputStatus;
  /** @nullable */
  note?: string | null;
}

export interface AttendanceDay {
  day: string;
  rate: number;
}

export interface LevelBreakdown {
  level: string;
  count: number;
}

export interface DashboardSummary {
  studentsCount: number;
  activeStudentsCount: number;
  teachersCount: number;
  attendanceRate: number;
  monthlyProgress: number;
  weeklyAttendance: AttendanceDay[];
  levelBreakdown: LevelBreakdown[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  student_joined: 'student_joined',
  attendance_recorded: 'attendance_recorded',
  progress_updated: 'progress_updated',
  teacher_joined: 'teacher_joined',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  createdAt: string;
}

export type GetActivityParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type ListStudentsParams = {
search?: string;
status?: ListStudentsStatus;
};

export type ListStudentsStatus = typeof ListStudentsStatus[keyof typeof ListStudentsStatus];


export const ListStudentsStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export type ListAttendanceParams = {
date?: string;
studentId?: number;
};