# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard summaries and activity
  - name: students
    description: Student management
  - name: teachers
    description: Teacher management
  - name: attendance
    description: Attendance tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Get recent activity
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, paused, completed]
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Student deleted
        "404":
          description: Student not found
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List teachers
      responses:
        "200":
          description: Teacher list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Teacher not found
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Attendance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: recordAttendance
      tags: [attendance]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Recorded attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        level:
          type: string
        progress:
          type: integer
        attendanceRate:
          type: integer
        status:
          type: string
          enum: [active, paused, completed]
        joinedAt:
          type: string
          format: date
      required: [id, name, email, level, progress, attendanceRate, status, joinedAt]
    StudentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        level:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
      required: [name, email, level]
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        level:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
    Teacher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        specialty:
          type: string
        studentsCount:
          type: integer
        status:
          type: string
          enum: [active, on_leave]
      required: [id, name, email, specialty, studentsCount, status]
    TeacherInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        specialty:
          type: string
        status:
          type: string
          enum: [active, on_leave]
      required: [name, email, specialty]
    TeacherUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        specialty:
          type: string
        status:
          type: string
          enum: [active, on_leave]
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        note:
          type: ["string", "null"]
      required: [id, studentId, studentName, date, status]
    AttendanceInput:
      type: object
      properties:
        studentId:
          type: integer
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        note:
          type: ["string", "null"]
      required: [studentId, date, status]
    DashboardSummary:
      type: object
      properties:
        studentsCount:
          type: integer
        activeStudentsCount:
          type: integer
        teachersCount:
          type: integer
        attendanceRate:
          type: integer
        monthlyProgress:
          type: integer
        weeklyAttendance:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceDay"
        levelBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/LevelBreakdown"
      required: [studentsCount, activeStudentsCount, teachersCount, attendanceRate, monthlyProgress, weeklyAttendance, levelBreakdown]
    AttendanceDay:
      type: object
      properties:
        day:
          type: string
        rate:
          type: integer
      required: [day, rate]
    LevelBreakdown:
      type: object
      properties:
        level:
          type: string
        count:
          type: integer
      required: [level, count]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [student_joined, attendance_recorded, progress_updated, teacher_joined]
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, description, createdAt]