# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceDay } from './attendanceDay';
import type { LevelBreakdown } from './levelBreakdown';

export interface DashboardSummary {
  studentsCount: number;
  activeStudentsCount: number;
  teachersCount: number;
  attendanceRate: number;
  monthlyProgress: number;
  weeklyAttendance: AttendanceDay[];
  levelBreakdown: LevelBreakdown[];
}