# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentStatus } from './studentStatus';

export interface Student {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  level: string;
  progress: number;
  attendanceRate: number;
  status: StudentStatus;
  joinedAt: Date;
}