# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentInputStatus } from './studentInputStatus';

export interface StudentInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  level: string;
  status?: StudentInputStatus;
}