# lib/api-zod/src/generated/types/teacher.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherStatus } from './teacherStatus';

export interface Teacher {
  id: number;
  name: string;
  email: string;
  specialty: string;
  studentsCount: number;
  status: TeacherStatus;
}