# lib/api-zod/src/generated/types/teacherInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherInputStatus } from './teacherInputStatus';

export interface TeacherInput {
  /** @minLength 2 */
  name: string;
  email: string;
  specialty: string;
  status?: TeacherInputStatus;
}